Dropped support for mpz accumulation with non-null initialization
We are currently not using it (it's not likely we will ever), and
initializing an mpz in the init function would put it in the wrong
memory context, crashing as soon as the aggregation needs to grow.

It looks like the agg memory context is not available in the init function.