More compact, versioned disk representation for mpq
Only the number of limbs in the numer is stored: the one in the denom
is calculated from the varlena header. This just take 28 bits: in the
remaining 4 bits store the version and the sign.

Use the trailing are reserved by the custom GMP allocators to avoid a
memcpy of the numer limbs. The denom limbs have to be copied anyway to
make the two limbs sequences contiguous.