Not exposing mpq(int, int) and mpq(numeric, numeric) anymore
They currently cause problems as mpq(n, d) with literal values may fail
if one argument fits into an int and one into a numeric.

It's easier to document that you can only pass small enough arguments
or use explicit mpz casts. mpq(numeric, numeric) is also ugly as it fails
on not integer args so I'm not entirely unhappy to hide it.